+++ src/langgraph_mcp/03_langgraph_agent_multiple_local_mcp.py
@@ -8,7 +8,7 @@ from typing import Annotated, List
 from langgraph.graph.message import add_messages
 from langgraph.checkpoint.memory import MemorySaver
 from langchain_mcp_adapters.client import MultiServerMCPClient
-from poetry_langgraph_mcp.configuration import get_llm
+from langgraph_mcp.configuration import get_llm
 
 """
 LangGraph ReAct Agent with Multiple MCP Servers
