+++ src/langgraph_mcp/05_langgraph_agent_multiple_external_mcp_fastapi.py
@@ -12,7 +12,7 @@ from typing import Annotated, List
 from langgraph.graph.message import add_messages
 from langgraph.checkpoint.memory import MemorySaver
 from langchain_mcp_adapters.client import MultiServerMCPClient
-from poetry_langgraph_mcp.configuration import get_llm
+from langgraph_mcp.configuration import get_llm
 
 VERBOSE = False
